+++ include/primitives.h
@@ -1,0 +1,8 @@
+/* primitives.h */
+
+#ifndef FILE_PRIMITIVES_SEEN
+#define FILE_PRIMITIVES_SEEN
+
+extern object *add_proc(object *arguments);
+
+#endif // FILE_PRIMITIVES_SEEN
